Extract useful functions into util file
